Fix(CI): Increase sleep time in workflow to fix race condition (#56)
The GitHub Actions workflow was failing with a `connection refused` error when trying to initialize the database. This indicates that the postgres service was not yet ready to accept connections.

This commit increases the sleep time in the `Initialize DB` step from 10 to 30 seconds to give the service more time to become available. This should make the workflow more robust and prevent this race condition from happening in the future.